Refactor optimization: rename bounds, decouple DEAP and config logic, achieve 100% coverage
